+++ modules/identity/main.tf
@@ -18,7 +18,7 @@ locals {
   ])
 
   groups                  = [for group in keys(var.memberships) : oci_identity_group.groups[group]]
-  service_accounts_groups = [for sa in var.service_accounts : oci_identity_group.service_accounts_groups[sa]]
+  service_accounts_groups = [for key, sa in var.service_accounts : oci_identity_group.service_accounts_groups[sa.name]]
 
   depends_on = concat(local.groups, local.service_accounts_groups)
 }
@@ -75,27 +75,40 @@ resource "oci_identity_user_group_membership" "user_group_membership" {
 # Service Accounts - to associate a policy with a service account, the
 # service account must belong to a group
 resource "oci_identity_user" "service_accounts" {
-  for_each = toset(var.service_accounts)
+  for_each = var.service_accounts
 
   compartment_id = var.tenant_id
-  description    = each.key
-  name           = each.key
+  description    = each.value.name
+  name           = each.value.name
 }
 
 resource "oci_identity_group" "service_accounts_groups" {
-  for_each = toset(var.service_accounts)
+  for_each = var.service_accounts
 
   compartment_id = var.tenant_id
-  description    = each.key
-  name           = each.key
+  description    = each.value.name
+  name           = each.value.name
 }
 
 resource "oci_identity_user_group_membership" "service_accounts_group_membership" {
-  for_each = toset(var.service_accounts)
+  for_each = var.service_accounts
 
-  group_id = oci_identity_group.service_accounts_groups[each.value].id
-  user_id  = oci_identity_user.service_accounts[each.value].id
+  group_id = oci_identity_group.service_accounts_groups[each.key].id
+  user_id  = oci_identity_user.service_accounts[each.key].id
 }
+
+resource "oci_identity_user_capabilities_management" "service_accounts_capabilities_management" {
+  for_each = var.service_accounts
+
+  user_id = oci_identity_user.service_accounts[each.key].id
+
+  can_use_api_keys             = lookup(each.value.capabilities, "api_keys", false)
+  can_use_auth_tokens          = lookup(each.value.capabilities, "auth_tokens", false)
+  can_use_console_password     = lookup(each.value.capabilities, "console_password", false)
+  can_use_customer_secret_keys = lookup(each.value.capabilities, "customer_secret_keys", false)
+  can_use_smtp_credentials     = lookup(each.value.capabilities, "smtp_credentials", false)
+}
+
 
 # Some policies have to be applied at the tenancy level so compartment_id must be the tenant_id
 resource "oci_identity_policy" "tenancy_policies" {